Different file for Registration View.
